requirements.txt
langgraph==0.2.61
requests==2.32.3
python-dotenv==1.0.1
fastapi==0.110.1
uvicorn==0.25.0
faiss-cpu==1.8.0
pydantic==2.7.4
pydantic-core==2.18.4
